Implemented tracking once an average passed a sections lower bound
This was done through using the checkDistribution function with a flag signifying it should check whether the lower bound of that distributions average for that range had been surpassed and if so track it. This is called every time a point was added.
Once it's noted that the lower bound has been surpassed then it imposes stricter conditions on distribution changes