Removed obsolete code
When I removed the dot on top of points from the visual side I didn't finish by removing from code as I wasn't sure if it was functional at the time.
Have reviewed the code and it was a purely visual element apart from for one point where I switched over to the more commonly used point_areas_list.
Now safe to purge the code of references to it.